Extract cleanup summary logger
